# docs/swagger.json
{"swagger":"2.0","info":{"contact":{"name":"PhantomX7","url":"https://github.com/PhantomX7","email":"zrphntm@gmail.com"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"paths":{"/post":{"post":{"description":"Create a new post with the provided details","consumes":["application/json"],"produces":["application/json"],"tags":["post"],"summary":"Create a new post","parameters":[{"description":"Post Create Request","name":"post","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.PostCreateRequest"}}],"responses":{"201":{"description":"Created","schema":{"allOf":[{"$ref":"#/definitions/response.Response"},{"type":"object","properties":{"data":{"$ref":"#/definitions/dto.PostResponse"}}}]}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/response.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/response.Response"}}}}},"/post/{id}":{"get":{"description":"Find a post with the provided ID","consumes":["application/json"],"produces":["application/json"],"tags":["post"],"summary":"Find a post by ID","parameters":[{"type":"integer","description":"Post ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"allOf":[{"$ref":"#/definitions/response.Response"},{"type":"object","properties":{"data":{"$ref":"#/definitions/dto.PostResponse"}}}]}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/response.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/response.Response"}}}},"put":{"description":"Update a post with the provided details","consumes":["application/json"],"produces":["application/json"],"tags":["post"],"summary":"Update a post","parameters":[{"type":"integer","description":"Post ID","name":"id","in":"path","required":true},{"description":"Post Update Request","name":"post","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.PostUpdateRequest"}}],"responses":{"200":{"description":"OK","schema":{"allOf":[{"$ref":"#/definitions/response.Response"},{"type":"object","properties":{"data":{"$ref":"#/definitions/dto.PostResponse"}}}]}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/response.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/response.Response"}}}},"delete":{"description":"Delete a post with the provided ID","consumes":["application/json"],"produces":["application/json"],"tags":["post"],"summary":"Delete a post","parameters":[{"type":"integer","description":"Post ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"allOf":[{"$ref":"#/definitions/response.Response"},{"type":"object","properties":{"data":{"type":"object"}}}]}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/response.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/response.Response"}}}}},"/user":{"post":{"description":"Create a new user with the provided details","consumes":["application/json"],"produces":["application/json"],"tags":["user"],"summary":"Create a new user","parameters":[{"description":"User Create Request","name":"user","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.UserCreateRequest"}}],"responses":{"201":{"description":"Created","schema":{"allOf":[{"$ref":"#/definitions/response.Response"},{"type":"object","properties":{"data":{"$ref":"#/definitions/dto.UserResponse"}}}]}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/response.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/response.Response"}}}}},"/user/{id}":{"get":{"description":"Find a user with the provided ID","consumes":["application/json"],"produces":["application/json"],"tags":["user"],"summary":"Find a user by ID","parameters":[{"type":"integer","description":"User ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"allOf":[{"$ref":"#/definitions/response.Response"},{"type":"object","properties":{"data":{"$ref":"#/definitions/dto.UserResponse"}}}]}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/response.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/response.Response"}}}},"put":{"description":"Update a user with the provided details","consumes":["application/json"],"produces":["application/json"],"tags":["user"],"summary":"Update a user","parameters":[{"type":"integer","description":"User ID","name":"id","in":"path","required":true},{"description":"User Update Request","name":"user","in":"body","required":true,"schema":{"$ref":"#/definitions/dto.UserUpdateRequest"}}],"responses":{"200":{"description":"OK","schema":{"allOf":[{"$ref":"#/definitions/response.Response"},{"type":"object","properties":{"data":{"$ref":"#/definitions/dto.UserResponse"}}}]}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/response.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/response.Response"}}}},"delete":{"description":"Delete a user with the provided ID","consumes":["application/json"],"produces":["application/json"],"tags":["user"],"summary":"Delete a user","parameters":[{"type":"integer","description":"User ID","name":"id","in":"path","required":true}],"responses":{"200":{"description":"OK","schema":{"allOf":[{"$ref":"#/definitions/response.Response"},{"type":"object","properties":{"data":{"type":"object"}}}]}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/response.Response"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/response.Response"}}}}}},"definitions":{"dto.PostCreateRequest":{"type":"object","required":["content","title"],"properties":{"content":{"type":"string"},"title":{"type":"string"}}},"dto.PostResponse":{"type":"object","properties":{"content":{"type":"string"},"created_at":{"type":"string"},"description":{"type":"string"},"id":{"type":"integer"},"title":{"type":"string"},"updated_at":{"type":"string"}}},"dto.PostUpdateRequest":{"type":"object","properties":{"content":{"type":"string"},"title":{"type":"string"}}},"dto.UserCreateRequest":{"type":"object","required":["name"],"properties":{"description":{"type":"string"},"name":{"type":"string"}}},"dto.UserResponse":{"type":"object","properties":{"created_at":{"type":"string"},"description":{"type":"string"},"id":{"type":"integer"},"name":{"type":"string"},"updated_at":{"type":"string"}}},"dto.UserUpdateRequest":{"type":"object","properties":{"description":{"type":"string"},"name":{"type":"string"}}},"response.Response":{"type":"object","properties":{"data":{},"error":{},"message":{"type":"string"},"meta":{},"status":{"type":"boolean"}}}}}